requirements.txt
entsoe-py>=0.6.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
matplotlib>=3.8.0
lightgbm>=4.3.0
xgboost>=2.0.0
scikit-learn>=1.4.0
seaborn>=0.13.0
joblib>=1.3.0
scipy>=1.12.0
autogluon.tabular>=1.1.0
google-generativeai>=0.8.0
flask>=3.0.0
apscheduler>=3.10.0
weasyprint>=62.0
jinja2>=3.1.0
openpyxl>=3.1.0
pytz>=2024.1
